commit finish

+++ src/main/java/pro/sky/calculator/CalculatorService.java
@@ -4,18 +4,18 @@ import org.springframework.stereotype.Service;
 
 @Service
 public class CalculatorService {
-    public int plus(int num1, int num2) {
-        return num1 + num2;
+    public int plus(double num1, double num2) {
+        return (int) (num1 + num2);
     }
 
-    public int minus(int num1, int num2) {
+    public double minus(double num1, double num2) {
         return num1 - num2;
     }
 
-    public int multiply(int num1, int num2) {
+    public double multiply(double num1, double num2) {
         return num1 * num2;
     }
-    public double divide(int num1, int num2) {
+    public double divide(double num1, double num2) {
         return (double) num1 / num2;
     }
 }
